requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose
bcrypt
passlib[bcrypt]
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
cx_Oracle
python-dotenv
